Fixes #76
As SLPreference does *not* inherit from Preference, we need to copy the public interface. This does not feel very nice, but it has been usual for a long time, so let's fix it again. Complements the changes in System-mt.1086.

Please note at some time we should consider to wrap or inherit from Preference.